Add files via upload
A desktop application has been developed using the WEKA library to obtain the appropriate data set content for each classification algorithm. For example; For the Naïve Bayes algorithm, the dataset must have only nominal data, for the Logistic Regression algorithm, the dataset must have only numerical data, for the K-Nearest Neighbor algorithm the dataset must be numeric. The data type is not important for Decision Tree algorithms. The dataset of the Artificial Neural Network algorithm must be digital. Finally, the data set for the Support Vector Machine algorithm must be numeric.